cp4ba_proddeploy_automation/requirements.txt
configobj
ruamel.yaml
pexpect
pandas
gitpython
pyodbc
requests
configparser
gitpython
pyyaml
yq
